+ File Save and Folder Creation
The GUI now has the ability to save a .JSON file, using the name provided by the user. A folder named "slideshows" will be created in the root project folder in order to store said save file. Currently the program saves this file, creates a slideshow folder if needed, and open and writes to the saved file.