Improved Importation(Now Works Somewhat)
Importing from a save file now displays the images used by the user. OldSaveInfo holds the images and their order, sounds and their order, image transitions, image transition intervals, whether the slideshow is manual or not, and the slideshow image interval length. All that needs to be done is to update the actual reels to show these changes from the previous slideshow, along with updating the exporter of this information.