ClickPopUp changes/Added Error Handling/Change to Exit Dialog
ClickPopUp now creates an option to change the selected image's transition time through the popup menu instead of having to change it by selecting the same transition type again.

The exit dialog has been reduced to the yes or no options instead of having a cancel option.

If the importation of the old directory used is unsuccessful, the program will present an error to the user before then closing.